refactor(meta): delete SST objects in meta node
