+++ src/meta/src/hummock/manager/tests.rs
@@ -2573,3 +2573,63 @@ async fn test_merge_compaction_group_task_expired() {
     // task 2 report failed due to the compaction group is merged
     assert!(!sst_ids.contains(&report_sst_id));
 }
+
+#[tokio::test]
+async fn test_vacuum() {
+    let (_env, hummock_manager, _cluster_manager, worker_id) = setup_compute_env(80).await;
+    let context_id = worker_id as _;
+    let hummock_meta_client: Arc<dyn HummockMetaClient> = Arc::new(MockHummockMetaClient::new(
+        hummock_manager.clone(),
+        context_id,
+    ));
+    assert_eq!(hummock_manager.delete_metadata().await.unwrap(), 0);
+    assert_eq!(hummock_manager.delete_object().await.unwrap(), 0);
+    hummock_manager.pin_version(context_id).await.unwrap();
+    let compaction_group_id = StaticCompactionGroupId::StateDefault.into();
+    let sst_infos = add_test_tables(
+        hummock_manager.as_ref(),
+        hummock_meta_client.clone(),
+        compaction_group_id,
+    )
+    .await;
+    assert_eq!(hummock_manager.delete_metadata().await.unwrap(), 0);
+    hummock_manager.create_version_checkpoint(1).await.unwrap();
+    assert_eq!(hummock_manager.delete_metadata().await.unwrap(), 6);
+    assert_eq!(hummock_manager.delete_metadata().await.unwrap(), 0);
+
+    assert!(hummock_manager.get_objects_to_delete().is_empty());
+    hummock_manager
+        .unpin_version_before(context_id, HummockVersionId::MAX)
+        .await
+        .unwrap();
+    hummock_manager.create_version_checkpoint(0).await.unwrap();
+    assert!(hummock_manager.get_objects_to_delete().is_empty());
+    hummock_manager
+        .complete_full_gc(
+            sst_infos
+                .iter()
+                .flat_map(|ssts| ssts.iter().map(|s| s.object_id))
+                .collect(),
+            None,
+            HashSet::default(),
+        )
+        .await
+        .unwrap();
+    assert_eq!(hummock_manager.get_objects_to_delete().len(), 3);
+    assert_eq!(
+        hummock_manager
+            .get_objects_to_delete()
+            .into_iter()
+            .sorted()
+            .collect::<Vec<_>>(),
+        sst_infos[0]
+            .iter()
+            .map(|s| s.object_id)
+            .sorted()
+            .collect::<Vec<_>>()
+    );
+    // No SST deletion is scheduled because no available worker.
+    assert_eq!(hummock_manager.delete_object().await.unwrap(), 3);
+    // No objects_to_delete.
+    assert_eq!(hummock_manager.delete_object().await.unwrap(), 0);
+}
